requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.23.2
motor
python-dotenv==1.0.0
pymongo==4.7.0
